Deparsing interval types
This one is a bit tricky. PG packs the interval parts into a single
integer and it does so according to a very specific set of combinations
enumerated in the grammar. There was no straightforward way to recompose
the original SQL - we have to use a map of known values.