Use actual Ruby classes for parser result, instead of decoded JSON
This is essentialy to enable easy and fast two-way communication with
the C library, and as a bonus makes for a better interaction on the Ruby
side, as we are handling actual objects instead of hashes and arrays.

Note this also unifies the scan output into the same protobuf file,
although the scan result retains a different top-level protobuf
message (ScanResult).