PgQuery.parse: Support complex queries with deeply nested ASTs
Requires at least google-protobuf v3.20.0, if using older releases this
change has no effect. To avoid complex dependency updates, we do not
require google-protobuf to be current when you use this library, so you
may want to pin it to the newer version if you depend on this behaviour.

Fixes #209