PgQuery::Node: Add inner and inner= helpers to modify inner object
Because of how oneof messages work in protobuf, its a bit tedious to
make modifications to a Node object, in particular when changing the type
of the inner object held within the Node in a generic code path (i.e.
with a dynamic inner type), which required first knowing the name of the
new inner object type, and then using public_send to set the new value.

To help, introduce the new "inner" and "inner=" methods for PgQuery::Node,
that get/set the inner object directly, avoiding the use of public_send.

This will be of particular help for anyone utilizing the walk! API to
make modifications to the query tree. To illustrate, an example is added
to the treewalker spec showing how to utilize this.