Update rubocop to 1.71, rubocop-rspec to 3.4
This also lets us drop the Ruby 3.0 requirement for rubocop on CI, since
we can now use rubocop on all released Ruby versions.

Further, turn off some of the newly introduced cops that seem to add
unnecessary noise, and opt in to new cops by default (this avoids a
warning, and seems fine since we lock the rubocop version).

Also, drop inline disabling of the "Lint/EmptyWhen" lint, which is no
longer needed.