Support null_test node type with parsing
When the query like SELECT col1 FROM t1 WHERE ((SELECT col2 FROM t2
LIMIT 1) IS NULL) is passed, the subselect_items contain the item with
the null_test node type. This null_test node type could contain a table
name or a function potentially, but that node type is ignored currently,
therefore these table names or functions weren't properly extracted as
tables/functions of the query.
With this change, by adding the argument of null_test node to the
subselect_items list, the arg will be evaluated again later on and will
be properly handled.